move copy_metadata function to image.tiff module
